Add Parallel routing to modals (#148)
* add parallet routing to new project modal

* cleanup

* modal naming relative to root

* parallel routing successfully working for projects edit and create
